thinkloop/requirements.txt
livekit-agents[google]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv>=0.19.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-multipart>=0.0.5
aiohttp>=3.8.1
asyncio>=3.4.3
pytz>=2021.3
supabase>=1.0.3
python-dateutil>=2.8.2 